requirements.txt
Django==4.1
psycopg==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
djangorestframework==3.13.1